+++ R/config.R
@@ -220,22 +220,36 @@ config_abort_type <- function(must_be, not, error_pkg, error_path, error_call) {
 
 config_abort <- function(pkg,
                          message,
+                         path = NULL,
                          ...,
                          call = caller_env(),
                          .envir = caller_env()) {
-
-  edit <- cli::format_inline("Edit {config_path(pkg)} to fix the problem.")
-
-  cli::cli_abort(
-    c(message, i = edit),
-    ...,
-    call = call,
-    .envir = .envir
-  )
+  
+  message <- config_message(pkg, message, path)
+  cli::cli_abort(message, ..., call = call, .envir = .envir)
 }
 
-config_path <- function(pkg) {
+config_warn <- function(pkg,
+                        message,
+                        path = NULL,
+                        ...,
+                        call = caller_env(),
+                        .envir = caller_env()) {
+  
+  message <- config_message(pkg, message, path)
+  cli::cli_warn(message, ..., call = call, .envir = .envir)
+}
+
+config_message <- function(pkg, message, path = NULL) {
   # Not all projects necessary have a _pkgdown.yml (#2542)
-  config <- pkgdown_config_path(pkg) %||% "_pkgdown.yml"
-  cli::style_hyperlink(path_file(config), paste0("file://", config))  
+  path <- path %||% pkgdown_config_path(pkg) %||% "_pkgdown.yml"
+  if (is_absolute_path(path)) {
+    path_label <- path_rel(path, pkg$src_path)
+  } else {
+    path_label <- path
+  }
+  
+  link <- cli::style_hyperlink(path_label, paste0("file://", path))  
+  message[[1]] <- paste0("In ", link, ", ", message[[1]])
+  message
 }